+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/Chargeable.java
@@ -2,7 +2,7 @@ package kenakata.catalog;
 
 public interface Chargeable
 {
-    double unitVat();
-    double unitCharge();
+    long unitVat();
+    long unitCharge();
     String label();
 }
